+++ QUANTAXIS/QAFetch/QATdx.py
@@ -37,7 +37,7 @@ from pytdx.exhq import TdxExHq_API
 from pytdx.hq import TdxHq_API
 from retrying import retry
 
-from QUANTAXIS.QAFetch.base import _select_market_code, _select_index_code, _select_type
+from QUANTAXIS.QAFetch.base import _select_market_code, _select_index_code, _select_type, _select_bond_market_code
 from QUANTAXIS.QAUtil import (QA_Setting, QA_util_date_stamp,
                               QA_util_date_str2int, QA_util_date_valid,
                               QA_util_get_real_date, QA_util_get_real_datelist,
@@ -723,7 +723,6 @@ def QA_fetch_get_bond_list(ip=None, port=None):
         return pd.concat([sz, sh], sort=False).query('sec=="bond_cn"').sort_index().assign(
             name=data['name'].apply(lambda x: str(x)[0:6]))
 
-
 @retry(stop_max_attempt_number=3, wait_random_min=50, wait_random_max=100)
 def QA_fetch_get_bond_day(code, start_date, end_date, frequence='day', ip=None,
                           port=None):
@@ -746,52 +745,11 @@ def QA_fetch_get_bond_day(code, start_date, end_date, frequence='day', ip=None,
         today_ = datetime.date.today()
         lens = QA_util_get_trade_gap(start_date, today_)
 
-        # sh
-        # 110×××
-        # 113xxxx
-
-        # sz
-        # 132xxxx
-        # 128xxxx
-
-        # 120×××企业债券；
-        # 129×××
-        # 100×××可转换债券；
-
-        # sz
-        # 10xxxx 国债现货
-        # 11xxxx 债券
-        # 12xxxx 可转换债券
-        # 12xxxx 国债回购
-
-        # if str(code)[0] in ['5', '1']:  # ETF
-        #     data = pd.concat([api.to_df(api.get_security_bars(
-        #         frequence, 1 if str(code)[0] in ['0', '8', '9', '5'] else 0,
-        #         code, (int(lens / 800) - i) * 800, 800))
-        #         for i in range(int(lens / 800) + 1)], axis=0, sort=False)
-        # else:
-        #     data = pd.concat([api.to_df(api.get_index_bars(
-        #         frequence, 1 if str(code)[0] in ['0', '8', '9', '5'] else 0,
-        #         code, (int(lens / 800) - i) * 800, 800))
-        #         for i in range(int(lens / 800) + 1)], axis=0, sort=False)
-
         code = str(code)
-
-        if code[0:3] in ['101', '104', '105', '106', '107', '108', '109',
-                         '111', '112', '114', '115', '116', '117', '118', '119',
-                         '123', '127', '128',
-                         '131', '139', ]:
-            # 深圳
-            data = pd.concat([api.to_df(api.get_security_bars(
-                frequence, 0,
-                code, (int(lens / 800) - i) * 800, 800))
-                for i in range(int(lens / 800) + 1)], axis=0, sort=False)
-        else:
-            data = pd.concat([api.to_df(api.get_security_bars(
-                frequence, 1,
-                code, (int(lens / 800) - i) * 800, 800))
-                for i in range(int(lens / 800) + 1)], axis=0, sort=False)
-
+        data = pd.concat([api.to_df(api.get_security_bars(
+            frequence, _select_bond_market_code(code),
+            code, (int(lens / 800) - i) * 800, 800))
+            for i in range(int(lens / 800) + 1)], axis=0, sort=False)
         data = data.assign(
             date=data['datetime'].apply(lambda x: str(x[0:10]))).assign(
             code=str(code)) \
@@ -831,16 +789,6 @@ def QA_fetch_get_bond_min(code, start, end, frequence='1min', ip=None,
     if lens > 20800:
         lens = 20800#u
     code = str(code)
-    def _select_bond_market_code(code):
-
-        if code[0:3] in ['101', '104', '105', '106', '107', '108', '109',
-                         '111', '112', '114', '115', '116', '117', '118', '119',
-                         '123', '127', '128',
-                         '131', '139', ]:
-            return 1
-        else:
-            return 0
-
     with api.connect(ip, port):
 
         data = pd.concat(
